Mohid Land modules - changes to be able to compile in Release single ModuleLagrangianGlobal - small verification related to Boom ModuleFillMatrix - replaced "where" expression in 3D fields in routine ConstructFillMatrix3D with SetMatrixVaue so that MOHID would not crash while debugging a large grid.
git-svn-id: https://mohid.svn.codeplex.com/svn@89409 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/MOHIDLand/ModuleBasin.F90
@@ -370,8 +370,8 @@ Module ModuleBasin
         real, dimension (:,:,:), pointer            :: Last5DaysAccRain => null ()
         real, dimension (:,:), pointer              :: Last5DaysAccRainTotal => null ()
         real, dimension (:,:), pointer              :: DailyAccRain => null ()
-        real, dimension (:,:), pointer              :: Current5DayAccRain => null ()
-        real, dimension (:,:), pointer              :: ActualCurveNumber => null ()
+        real(8), dimension (:,:), pointer           :: Current5DayAccRain => null ()
+        real(8), dimension (:,:), pointer           :: ActualCurveNumber => null ()
         real, dimension (:,:), pointer              :: S => null ()
         type (T_Time)                               :: NextRainAccStart 
         real                                        :: CIDormThreshold = 12.70,         &
